fix: automatically run "go mod tidy" if no packages are returned
For some reason `packages.Load()` returns no errors and an empty package
list if `go mod tidy` needs to be run. The underlying `go list` is
returning this:

```
$ go list .
go: updates to go.mod needed; to update it:
	go mod tidy
```

We can't be sure that this will fix the issue, but trying it can't hurt,
and if the packages load succesfully afterwards, that's one less thing
the user needs to worry about.